+++ src/tests/IntegrationTests/Memory/SafeMemoryStreamTests.cs
@@ -45,6 +45,20 @@ namespace SafeOrbit.Memory
             Assert.AreEqual(eofResult, -1);
         }
 
+        [Test]
+        public void Can_write_read_and_close()
+        {
+            Assert.DoesNotThrow(() =>
+            {
+                var expected = new byte[] { 5, 10, 15, 20, 25, 30 };
+                using var sut = new SafeMemoryStream();
+                var buffer = expected.CopyToNewArray();
+                sut.Write(buffer, 0, buffer.Length);
+                var part1 = new byte[buffer.Length];
+                sut.Read(part1, 0, 3);
+            });
+        }
+
         protected override SafeMemoryStream GetSut()
             => new SafeMemoryStream();
     }