+++ src/tests/UnitTests/Memory/SafeMemoryStreamTests.cs
@@ -1,4 +1,5 @@
-﻿using NUnit.Framework;
+﻿using System;
+using NUnit.Framework;
 using SafeOrbit.Extensions;
 
 namespace SafeOrbit.Memory
@@ -8,16 +9,73 @@ namespace SafeOrbit.Memory
     public class SafeMemoryStreamTests
     {
         [Test]
-        public void Write_Clears_Input_Buffer()
+        public void Ctor_GivenBuffer_Writes()
         {
-            using var sut = GetSut();
-            var buffer = new byte[] {5, 10, 15, 20, 25, 30};
-            sut.Write(buffer, 0, buffer.Length);
+            // Arrange
+            var expected = new byte[] { 5, 10, 15, 20, 25, 30 };
+            var input = expected.CopyToNewArray();
+
+            // Act
+            var sut = new SafeMemoryStream(input);
+
+            // Assert
+            var actual = new byte[expected.Length];
+            sut.Read(actual, 0, expected.Length);
+            CollectionAssert.AreEqual(expected, actual);
+        }
+
+        [Test]
+        public void Ctor_GivenBuffer_Clears()
+        {
+            // Arrange
+            var buffer = new byte[] { 5, 10, 15, 20, 25, 30 };
+
+            // Act
+            _ = new SafeMemoryStream(buffer);
+
+            // Assert
             Assert.That(buffer, Is.Empty.Or.All.EqualTo(0));
         }
 
         [Test]
-        public void Read_Clears_Inner_Buffer()
+        public void Write_GivenBuffer_Clears()
+        {
+            // Arrange
+            using var sut = GetSut();
+            var buffer = new byte[] {5, 10, 15, 20, 25, 30};
+
+            // Act
+            sut.Write(buffer, 0, buffer.Length);
+
+            // Assert
+            Assert.That(buffer, Is.Empty.Or.All.EqualTo(0));
+        }
+
+        [Test]
+        public void Write_AfterBeingRead_CanWrite()
+        {
+            // Arrange
+            var expected1 = new byte[] { 5, 10, 15 };
+            var expected2 = new byte[] { 20, 25, 30 };
+            using var sut = GetSut();
+            var buffer = expected1.CopyToNewArray();
+            sut.Write(buffer, 0, buffer.Length);
+            var actual1 = new byte[expected1.Length];
+            sut.Read(actual1, 0, 3);
+
+            // Act
+            buffer = expected2.CopyToNewArray();
+            sut.Write(buffer, 0, buffer.Length);
+            var actual2 = new byte[expected2.Length];
+            sut.Read(actual2, 0, 3);
+
+            // Assert
+            Assert.That(actual1, Is.EqualTo(expected1));
+            Assert.That(actual2, Is.EqualTo(expected2));
+        }
+
+        [Test]
+        public void Read_WrittenInnerBuffer_Clears() // a.k.a. FIFO stream
         {
             // Arrange
             var expected1 = new byte[] {5, 10, 15};
@@ -36,29 +94,6 @@ namespace SafeOrbit.Memory
             Assert.That(part1, Is.EqualTo(expected1));
             Assert.That(part2, Is.EqualTo(expected2));
             Assert.That(sut.Length, Is.Zero);
-        }
-
-        [Test]
-        public void Write_AfterBeingRead_CanWrite()
-        {
-            //arrange
-            var expected1 = new byte[] {5, 10, 15};
-            var expected2 = new byte[] {20, 25, 30};
-            using var sut = GetSut();
-            var buffer = expected1.CopyToNewArray();
-            sut.Write(buffer, 0, buffer.Length);
-            var actual1 = new byte[expected1.Length];
-            sut.Read(actual1, 0, 3);
-
-            // Act
-            buffer = expected2.CopyToNewArray();
-            sut.Write(buffer, 0, buffer.Length);
-            var actual2 = new byte[expected2.Length];
-            sut.Read(actual2, 0, 3);
-
-            // Assert
-            Assert.That(actual1, Is.EqualTo(expected1));
-            Assert.That(actual2, Is.EqualTo(expected2));
         }
 
         [Test]